requirements.txt
altgraph==0.17.3
colorama==0.4.6
comtypes==1.1.14
docx2pdf==0.1.8
et-xmlfile==1.1.0
lxml==4.9.2
numpy==1.24.2
openpyxl==3.1.1
pandas==1.5.3
pefile==2023.2.7
pyinstaller==5.1
pyinstaller-hooks-contrib==2023.0
pypandoc==1.10
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7.1
pywin32==305
pywin32-ctypes==0.2.0
six==1.16.0
tqdm==4.64.1
